Rename EnclaveMediator to TransactionManager.
