+++ jaxrs-service/src/test/java/com/quorum/tessera/jaxrs/JaxrsITConfig.java
@@ -5,7 +5,7 @@ import com.quorum.tessera.config.Config;
 import com.quorum.tessera.config.JdbcConfig;
 import com.quorum.tessera.config.KeyConfiguration;
 import com.quorum.tessera.config.ServerConfig;
-import com.quorum.tessera.enclave.EnclaveMediator;
+import com.quorum.tessera.transaction.TransactionManagerImpl;
 import com.quorum.tessera.node.PartyInfoParser;
 import com.quorum.tessera.node.PartyInfoService;
 import java.nio.file.Path;
@@ -20,8 +20,8 @@ import org.springframework.context.annotation.ImportResource;
 public class JaxrsITConfig {
     
     @Bean(name = "enclaveDelegate")    
-    public EnclaveMediator enclaveMediator() {
-        return mock(EnclaveMediator.class);
+    public TransactionManagerImpl enclaveMediator() {
+        return mock(TransactionManagerImpl.class);
     }
     
     @Bean
